added board/post auth + post comments

+++ base/views.py
@@ -1,9 +1,10 @@
 from django.shortcuts import render, redirect
-from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
-from .models import Board, Post, User
+from .models import Board, Post, User, Comment
 from .forms import BoardForm, PostForm, RegisterUserForm
 
 
@@ -54,7 +55,6 @@ def register_user(request):
     content={'form':form}
     return render(request,'base/register.html',content)
 
-
 #board
 def board(request, pk):
     target_board = Board.objects.get(topic=pk)
@@ -63,20 +63,30 @@ def board(request, pk):
     context = {'board':target_board,'posts':posts}
     return render(request,'base/board.html',context)
 
+@login_required(login_url='login')
 def create_board(request):
     form = BoardForm()
+
     if request.method == 'POST': #post handling
         form = BoardForm(request.POST)
         if form.is_valid():
-            form.save()
+            board = form.save(commit=False)
+            board.host = request.user
+            board.topic = board.topic.lower()
+            board.save()
             return redirect('home')
 
     context={'form':form}
     return render(request,'base/board-form.html',context)
 
+@login_required(login_url='login')
 def edit_board(request,pk):
     board = Board.objects.get(topic=pk)
     form = BoardForm(instance=board)
+
+    if request.user != board.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         form=BoardForm(request.POST, instance=board)
         if form.is_valid():
@@ -86,8 +96,13 @@ def edit_board(request,pk):
     context={'form':form}
     return render(request,'base/board-form.html',context)
 
+@login_required(login_url='login')
 def delete_board(request,pk):
     board = Board.objects.get(topic=pk)
+
+    if request.user != board.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         board.delete()
         return redirect('home')
@@ -96,13 +111,26 @@ def delete_board(request,pk):
 #post
 def post(request, pk, pk2):
     post = Post.objects.get(id=pk2)
+    comments = post.comment_set.all().order_by('-created')
+
+    #comment posting
+    if request.method == "POST":
+        comment = Comment.objects.create(
+            creator = request.user,
+            post = post,
+            content = request.POST.get('content')
+        )
+        return redirect('post',pk=pk,pk2=pk2)
+
+
     title=pk
-    context={'post':post,'title':title}
+    context={'post':post,'title':title,'comments':comments}
     return render(request,'base/post.html',context)
 
+@login_required(login_url='login')
 def create_post(request,pk):
     form = PostForm()
-    user = User.objects.get(id=1) #change when users added in
+    user = User.objects.get(id=request.user.id) #change when users added in
     board = Board.objects.get(topic=pk)
     title=pk
 
@@ -119,10 +147,14 @@ def create_post(request,pk):
     context={'form':form,'title':title}
     return render(request,'base/post-form.html',context)
 
+@login_required(login_url='login')
 def edit_post(request, pk, pk2):
     post = Post.objects.get(id=pk2)
     form = PostForm(instance=post)
     title=pk
+
+    if request.user != post.creator:
+        return HttpResponse('You are not allowed here')
 
     if request.method == 'POST':
         form=PostForm(request.POST, instance=post)
@@ -133,9 +165,29 @@ def edit_post(request, pk, pk2):
     context={'form':form,'title':title}
     return render(request,'base/post-form.html',context)
 
+@login_required(login_url='login')
 def delete_post(request, pk, pk2):
     post = Post.objects.get(id=pk2)
+
+    if request.user != post.creator:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         post.delete()
         return redirect('board',pk=pk)
     return render(request, 'base/delete.html', {'obj':post})
+
+#comment
+
+@login_required(login_url='login')
+def delete_comment(request,pk,pk2,pk3):
+    comment = Comment.objects.get(id=pk3)
+
+    if request.user != comment.creator:
+        return HttpResponse('You are not allowed here')
+
+    if request.method == "POST":
+        comment.delete()
+        return redirect('post',pk=pk,pk2=pk2)
+    
+    return render(request,'base/delete.html',{'obj':comment})
